% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters_joins.R
\name{join_sp_to_tmhmm}
\alias{join_sp_to_tmhmm}
\title{Combine SignalP and TMHMM results}
\usage{
join_sp_to_tmhmm(tmhmm, sp, is.secreted = TRUE)
}
\arguments{
\item{tmhmm}{Tibble from read_tmhmm()}

\item{sp}{Tibble from read_sp3(), sp5 not yet supported.}

\item{is.secreted}{Also filter the results down to only secretions.}
}
\description{
Further, filter results on whether or not proteins are secreted.
This means right now, TMHMM results must be more comprehensive
than SignalP results.
}
\seealso{
\code{/link{read_tmhmm}}
}
