% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_inputs.R
\name{pivot_tmhmm}
\alias{pivot_tmhmm}
\title{Tidy TMHMM-2.0c data}
\usage{
pivot_tmhmm(x)
}
\arguments{
\item{x}{tab-separated TMHMM-2.0c output}
}
\value{
A tibble
}
\description{
After reading TMHMM-2.0c output into R, it's still in an unfriendly-format.
This tidies the "topology" string in the last column of the output
}
\examples{
tmhmm_test <- tibble(Protein = c("A", "B", "C", "D", "E"),
                     Info = c(1, 2, 3, 4, 5),
                     Topology = c("o", "i", "i4-9",
                                  "o10-20i30-50",
                                  "i105-205o305-405i505-605"))

pivot_tmhmm(tmhmm_test)

# This can now be used in ggplot
# e.g. How much residue length is TM-helix folds?

pivot_tmhmm(tmhmm_test) \%>\%
  mutate(local_in = Stop-Start+1) \%>\%
  group_by(Protein) \%>\%
  mutate(Total_in = sum(local_in, na.rm = TRUE)) \%>\%
  distinct(Protein, .keep_all = TRUE) \%>\%
  select(-Local_in) \%>\%
  ggplot2::ggplot(.) +
  ggplot2::geom_histogram(aes(Total_in))
}
\seealso{
\code{/link{read_tmhmm}} for reading input with proper names.
}
\keyword{tmhmm}
