% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters_joins.R
\name{prune_tm}
\alias{prune_tm}
\title{Filter proteins on TMHMM results}
\usage{
prune_tm(x)
}
\arguments{
\item{x}{Tibble from internally joined data}
}
\description{
Removes proteins from joined TMHMM-SignalP data
}
