% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_inputs.R
\name{read_sp3}
\alias{read_sp3}
\title{Process SignalP-3.1 output}
\usage{
read_sp3(file)
}
\arguments{
\item{file}{tab-separated output file from SignalP-3.1}
}
\value{
A tibble
}
\description{
Read in a file from SignalP-3.1.
Right now, this only handles the hmm setting
This function simply reads in the file and adds logical column names.
}
\keyword{signalp}
