% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_inputs.R
\name{read_sp5}
\alias{read_sp5}
\title{Process SignalP-5 output}
\usage{
read_sp5(file)
}
\arguments{
\item{file}{gff output file from SignalP-5}
}
\value{
A tibble
}
\description{
Read in a file from SignalP-5.
Right now, it only handles gff output (not the default SignalP-5 setting).
This function simply reads in the file and adds logical column names.
}
\keyword{signalp}
