% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_inputs.R
\name{read_tmhmm}
\alias{read_tmhmm}
\title{Process TMHMM-2.0c output}
\usage{
read_tmhmm(
  file,
  cols = c("Protein", "Length", "Exp_AA", "First_60", "Pred_hel", "Topology")
)
}
\arguments{
\item{file}{gff output file from SignalP-5}
}
\value{
A tibble
}
\description{
Read in a file from TMHMM-2.0c.
This function simply reads in the file and adds logical column names
}
\seealso{
\code{/link{pivot_tmhmm}} for conversion to "long" format.
}
\keyword{tmhmm}
